Initializes page with style and layout
With this commit, most of the layout of the html containers should be
done. There might be small tweaks that need ot be done. The basic
layout of the page also is done. To make the layout of the page more
obvious, I added dummy colors to the page. I need to go back and
confirm the correct colors and font properties are used.