Sanitizes and solitifies header
I realized that I wanted to just make a generic page, so I replaced the
custom name for the business with generic content.

This change also fixes a lot of the spacing issues. The header will also
now grow or shrink with the page up to an extent so that it fills
most browser pages. There still needs to be a mobile version of
the page, but that is to be done later.