Sanitizes and solitifies card section
I realized that I wanted to just make a generic page, so I replaced the
custom content for the business with generic content. This includes
dummy images from a dummy image website.

Like the header changes, this fixes spacing problems. This section also
now grows or shrinks to fill the screen.