Makes it look better on mobile
I changed several sections to wrap like the title section,
the card section, and the sign up section.

I also moved a bunch of the padding and margin around
so that it looked better on mobile. I still need to make a
hamburger menu for the links in the header since they
can't fit on the mobile view too well.

The footer copyright notice was also too big, so I reduced
the font size.